JAVA PROJECT 1
Menu that offers several different types of operations.  Finding Product
of positive numbers user enters (ignores negatives), Finding the index
of the first and last occurrence of the number 12 in a string of numbers
the user enters,  Finding Max Min Avg of a string of numbers the user
enters, Finding the grade value of the scores entered by the user also
summarizes the number of As Bs Cs ... that have occurred in that string
of numbers the user enters plus gives the Avg Max Min as well.